# contracts/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Lantern API
  description: |
    Canonical public API contract for the Lantern anonymous social client.

    All public content is authored by an anonymous persona. Persona objects never
    include email addresses, real-profile identifiers, or internal moderation notes.
    Real-profile data is available only through `/v1/me` to an authenticated session.
  version: "1.0.0"
  contact:
    name: Lantern Rebuild Team
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: "{scheme}://{host}:{port}"
    description: Local development server (default)
    variables:
      scheme:
        enum:
          - http
          - https
        default: http
      host:
        default: localhost
      port:
        default: "8080"
  - url: https://api-staging.lantern.app
    description: Staging server
  - url: https://api.lantern.app
    description: Production server
security:
  - BearerAuth: []
...
paths:
...
  /v1/auth/email-codes:
    post:
      operationId: sendEmailCode
      summary: Request an email verification code
      description: |
        Sends a one-time verification code to the supplied email address. The code is
        used to create or resume a session. Rate-limited per email address.
      security: []
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailCodeRequest"
      responses:
        "204":
          description: Verification code sent.
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "429":
          $ref: "#/components/responses/RateLimited"
  /v1/auth/email-sessions:
    post:
      operationId: createEmailSession
      summary: Exchange a verification code for a session
      description: |
        Verifies the code and returns an access/refresh token pair. Creates a new real
        profile if the email has never been used.
      security: []
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmailSessionRequest"
      responses:
        "201":
          description: Session created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
  /v1/auth/refresh:
    post:
      operationId: refreshSession
      summary: Refresh an access token
      description: |
        Exchanges a valid refresh token for a new access token. The old access token
        may be rejected shortly after this call.
      security: []
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "200":
          description: New access token issued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessTokenResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/auth/session:
    delete:
      operationId: deleteSession
      summary: Revoke the current session
      description: Revokes the access token used to authenticate the request.
      tags:
        - Auth
      responses:
        "204":
          description: Session revoked.
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/me:
    get:
      operationId: getMe
      summary: Get the authenticated real profile
      description: |
        Returns the authenticated real profile, including masked email and persona
        management metadata. Public content is never returned here.
      tags:
        - Identity
      responses:
        "200":
          description: Real profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RealProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      operationId: deleteMe
      summary: Request account deletion
      description: |
        Starts the 30-day account deletion grace period. Requires a verification code
        sent to the real profile email. Within the grace period, logging in cancels the
        deletion.
      tags:
        - Identity
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountDeletionRequest"
      responses:
        "200":
          description: Deletion initiated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountDeletionState"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer access token returned by `POST /v1/auth/email-sessions` or `POST /v1/auth/refresh`.
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Client-generated unique key for safely retrying mutating requests. The server
        stores the response for at least 24 hours and replays it for duplicate keys
        targeting the same endpoint and authenticated principal.
      schema:
        type: string
        minLength: 8
        maxLength: 128
  schemas:
...
    ErrorCode:
      type: string
      description: Stable machine-readable error code.
      enum:
        - unauthorized
        - forbidden
        - not_found
        - method_not_allowed
        - conflict
        - rate_limited
        - internal_error
        - not_implemented
        - validation.failed
        - idempotency.conflict
        - idempotency.missing_key
        - auth.invalid_email
        - auth.invalid_code
        - auth.rate_limited
        - auth.session_expired
        - auth.session_revoked
        - auth.invalid_token
        - me.not_found
        - me.deletion_invalid_code
        - me.deletion_already_pending
        - me.email_change_invalid_code
        - me.email_already_used
        - me.email_change_rate_limited
        - me.export_rate_limited
        - me.export_not_ready
        - me.export_not_found
        - persona.not_found
        - persona.not_owned
        - persona.max_reached
        - persona.alias_taken
        - persona.alias_disallowed
        - persona.archived
        - persona.restricted
        - persona.default_required
        - validation.export_format_invalid
        - topic.not_found
        - topic.already_followed
        - topic.not_followed
        - topic.hidden
        - post.not_found
        - post.not_author
        - post.invalid_state
        - post.content_disallowed
        - post.topic_required
        - post.rate_limited
        - comment.not_found
        - comment.not_author
        - comment.invalid_state
        - comment.content_disallowed
        - comment.parent_not_found
        - comment.rate_limited
        - reaction.invalid_type
        - reaction.already_exists
        - reaction.not_found
        - reaction.target_not_found
        - save.already_exists
        - save.not_found
        - media.invalid_format
        - media.too_large
        - block.not_found
        - block.already_exists
        - block.self
        - report.invalid_target
        - report.target_not_found
        - report.duplicate
        - report.details_required
        - report.self
        - moderation.case_not_found
        - moderation.invalid_outcome
        - moderation.invalid_status
        - moderation.report_not_found
        - moderation.not_moderator
        - moderation.rate_limited
    Error:
      type: object
      required:
        - code
        - message
        - requestId
      properties:
        code:
          $ref: "#/components/schemas/ErrorCode"
        message:
          type: string
          description: Human-readable description of the error.
        requestId:
          type: string
          format: uuid
          description: Request ID to include when contacting support.
    User:
      type: object
      required:
        - id
        - emailMasked
        - status
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        emailMasked:
          type: string
          description: Masked email address, e.g. `a•••@example.com`.
          example: a•••@example.com
        status:
          type: string
          enum:
            - active
            - deleting
            - deleted
            - suspended
            - banned
        createdAt:
          type: string
          format: date-time
    RealProfile:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          required:
            - updatedAt
            - defaultPersonaId
            - personaCount
            - maxPersonas
          properties:
            updatedAt:
              type: string
              format: date-time
            defaultPersonaId:
              type: string
              format: uuid
              nullable: true
            personaCount:
              type: integer
              minimum: 0
            maxPersonas:
              type: integer
              minimum: 1
              example: 5
            deletion:
              type: object
              nullable: true
              allOf:
                - $ref: "#/components/schemas/AccountDeletionState"
    Session:
      type: object
      required:
        - accessToken
        - refreshToken
        - tokenType
        - expiresIn
        - personaId
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
          enum:
            - Bearer
        expiresIn:
          type: integer
          description: Access token lifetime in seconds.
        personaId:
          type: string
          format: uuid
          nullable: true
          description: Default persona if one has been set.
        isStaff:
          type: boolean
          description: True when the authenticated account has staff privileges.
    AccessTokenResponse:
      type: object
      required:
        - accessToken
        - tokenType
        - expiresIn
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
          enum:
            - Bearer
        expiresIn:
          type: integer
        isStaff:
          type: boolean
          description: True when the authenticated account has staff privileges.
...
    EmailCodeRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        purpose:
          type: string
          enum:
            - login
            - email_change
            - deletion
          default: login
          description: Purpose of the verification code.
    EmailSessionRequest:
      type: object
      required:
        - email
        - code
      properties:
        email:
          type: string
          format: email
          maxLength: 254
        code:
          type: string
          pattern: "^[0-9]{6}$"
    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
    AccountDeletionRequest:
      type: object
      required:
        - verificationCode
      properties:
        verificationCode:
          type: string
          pattern: "^[0-9]{6}$"
    AccountDeletionState:
      type: object
      required:
        - gracePeriodEndsAt
      properties:
        gracePeriodEndsAt:
          type: string
          format: date-time
...
  responses:
    BadRequest:
      description: The request was malformed or failed basic validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication is required or the access token is invalid/expired.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: The authenticated principal lacks permission for the resource.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Conflict:
      description: The request conflicts with the current state (duplicate, idempotency mismatch, etc.).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    UnprocessableEntity:
      description: The request was syntactically valid but semantically invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    RateLimited:
      description: Too many requests. Retry after the period indicated by `Retry-After`.
      headers:
        Retry-After:
          description: Seconds until the client may retry.
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"